+++ changelog.md
@@ -28,3 +28,7 @@
 ### 2017-03-25
 [Mogwai]
 * weenies weenies everywhere. object structure pass 1
+
+### 2017-03-30
+[StackOverflow]
+* Start of a SpellFactory + World Object Fixes and a real Vector3 - AceVector